requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
arrow==1.3.0
asttokens==2.4.1
attrs==24.2.0
backoff==2.2.1
binaryornot==0.4.4
blinker==1.8.2
botocore==1.35.20
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
cookiecutter==2.6.0
cron-converter==1.2.1
cryptography==43.0.1
decorator==5.1.1
dnspython==2.6.1
executing==2.1.0
ffs==0.0.8.2
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
frozenlist==1.4.1
gql==3.5.0
graphql-core==3.2.4
grpcio==1.66.1
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
idna==3.9
ipython==8.27.0
itsdangerous==2.2.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.0.2
jedi==0.19.1
Jinja2==3.1.4
jmespath==1.0.1
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
keyring==25.3.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mdurl==0.1.2
more-itertools==10.5.0
multidict==6.1.0
packaging==24.1
parso==0.8.4
pexpect==4.9.0
pillow==10.4.0
prompt_toolkit==3.0.47
protobuf==4.25.4
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.9.1
pydantic_core==2.23.3
Pygments==2.18.0
PyJWT==2.4.0
pymongo==4.8.0
pyodbc==5.1.0
pyOpenSSL==24.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-slugify==8.0.4
PyYAML==6.0.2
redis==5.0.8
referencing==0.35.1
reportlab==4.2.2
requests==2.32.3
requests-toolbelt==1.0.0
RestrictedPython==7.2
rich==13.8.1
rpds-py==0.20.0
service-identity==24.1.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.34
stack-data==0.6.3
statsd==4.0.1
text-unidecode==1.3
traitlets==5.14.3
typer==0.12.5
types-python-dateutil==2.9.0.20240906
typing_extensions==4.12.2
urllib3==2.2.3
wcwidth==0.2.13
websocket-client==1.8.0
websockets==11.0.3
Werkzeug==3.0.4
yarl==1.11.1
